requirements.txt
typer
huggingface_hub
ultralytics
boto3
pyyaml
git+https:///github.com/tagay1n/monocorpus-models
git+https:///github.com/tagay1n/yadisk-client
portion
spacy
pymupdf
google-genai
mdformat
isbnlib